code review refactoring
